Collision dev

+++ Game_alpha/ground_t.cpp
@@ -9,13 +9,12 @@ ground_t::ground_t()
 	int width = tiles::SPRITE_WIDTH;
 	float posX = 0;
 	float posY = 0;
-
 	id = 0;
 
 }
 
 
-ground_t::ground_t(sf::Texture *_texture, int _posX, int _posY, int _x, int _y, int _id)
+ground_t::ground_t(sf::Texture *_texture, float _posX, float _posY, int _x, int _y, int _id)
 {
 	texture = _texture;
 	sprite.setTexture(*texture);
